Make sure we clean up in solveHelper
Since eachSolution clients sometimes use exceptions for flow control,
we need to make sure that the board ends up in the same state before and
after eachSolution is called, no matter what.  A better approach would
be to restore during the top level call to solveHelper only if an
exception has occurred, or better yet, to not use exceptions for flow
control.